remove __SSE4_2__ support

+++ cityhash/city.cc
@@ -27,7 +27,6 @@
 // possible hash functions, by using SIMD instructions, or by
 // compromising on hash quality.
 
-#include "config.h"
 #include "city.h"
 
 #include <algorithm>
@@ -143,7 +142,7 @@ static uint32 Hash32Len13to24(const char *s, size_t len) {
 static uint32 Hash32Len0to4(const char *s, size_t len) {
   uint32 b = 0;
   uint32 c = 9;
-  for (int i = 0; i < len; i++) {
+  for (size_t i = 0; i < len; i++) {
     b = b * c1 + s[i];
     c ^= b;
   }
@@ -485,7 +484,7 @@ uint128 CityHash128(const char *s, size_t len) {
 }
 
 #ifdef __SSE4_2__
-#include <citycrc.h>
+#include "citycrc.h"
 #include <nmmintrin.h>
 
 // Requires len >= 240.
